+++ src/components/interview/RightBubble.tsx
